fix: deleted duplicated files users
