+++ src/backend/routes/users/routes.ts
@@ -1,53 +1,15 @@
 // THESE ARE THE ROUTES FOR USERS
-import express from "express";
-import { PrismaClient } from "@prisma/client";
 
-const router = express.Router();
-const prisma = new PrismaClient();
+import { Request, Response, Router } from "express";
+import { UserController } from "./controllers.js";
+import { UserData } from "./models.js";
 
-// GET route to fetch all user profiles
-router.get("/api/users", async (req, res) => {
-  try {
-    const users = await prisma.userProfile.findMany();
-    res.json(users);
-  } catch (error) {
-    console.error("Error fetching users:", error);
-    res.status(500).json({ error: "Failed to fetch users" });
-  }
+const router = Router();
+
+// Define routes for users
+router.post("/", (req: Request<UserData>, res: Response<UserData>) => {
+  UserController.createUser(req, res);
 });
-
-// Function to create a new user
-router.post("/api/users", async (req, res) => {
-  try {
-    const userData = req.body;
-    const newUser = await createUser(userData);
-
-    res.status(201).json(newUser);
-  } catch (err) {
-    console.error("Error creating user", err);
-    res.status(500).json({ error: "Failed to create user" });
-  }
-});
-
-const createUser = async (userData: any) => {
-  try {
-    const newUser = await prisma.userProfile.create({
-      data: {
-        name: userData.name,
-        surname: userData.surname,
-        avatarImage: userData.avatarImage,
-        email: userData.email,
-        password: userData.password,
-        birthDate: userData.birthDate,
-        primaryAddress: userData.primaryAddress,
-        secondaryAddress: userData.secondaryAddress,
-      },
-    });
-    return newUser;
-  } catch (error) {
-    console.error("Error creating user:", error);
-    throw new Error("Failed to create user");
-  }
-};
+// Add more routes as needed
 
 export default router;
