+++ tsconfig.json
@@ -22,7 +22,7 @@
       "@/*": ["./src/*"]
     },
   },
-  "include": ["next-env.d.ts", "src/**/*.ts", "src/**/*.tsx", ".next/types/**/*.ts", "src/backend/index.mjs", "src/backend/users/models.ts", "src/backend/users/controllers.ts", "src/backend/users/routes.ts"],
+  "include": ["next-env.d.ts", "src/**/*.ts", "src/**/*.tsx", ".next/types/**/*.ts", "src/backend/index.mjs", "src/backend/routes/users/models.ts", "src/backend/routes/users/controllers.ts", "src/backend/routes/users/routes.ts"],
   "exclude": ["node_modules"],
 }
